# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PREXCOL API",
        "description": "API documentation for PREXCOL",
        "termsOfService": "https://www.example.com/terms/",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "BSD License"
        },
        "version": "v1"
    },
    "basePath": "/api",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "paths": {
...
        "/productos/productos/{id}/": {
            "get": {
                "operationId": "productos_productos_read",
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Producto"
                        }
                    }
                },
                "tags": [
                    "productos"
                ]
            },
            "put": {
                "operationId": "productos_productos_update",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Producto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Producto"
                        }
                    }
                },
                "tags": [
                    "productos"
                ]
            },
            "patch": {
                "operationId": "productos_productos_partial_update",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Producto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Producto"
                        }
                    }
                },
                "tags": [
                    "productos"
                ]
            },
            "delete": {
                "operationId": "productos_productos_delete",
                "description": "",
                "parameters": [],
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "tags": [
                    "productos"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ]
        },
...
    },
    "definitions": {
...
        "Producto": {
            "required": [
                "nombre",
                "descripcion",
                "precio",
                "tienda",
                "proveedor"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "tienda_nombre": {
                    "title": "Tienda nombre",
                    "type": "string",
                    "readOnly": true
                },
                "proveedor_nombre": {
                    "title": "Proveedor nombre",
                    "type": "string",
                    "readOnly": true
                },
                "secciones": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "uniqueItems": true
                },
                "secciones_nombres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "readOnly": true,
                    "uniqueItems": true
                },
                "nombre": {
                    "title": "Nombre",
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "descripcion": {
                    "title": "Descripcion",
                    "type": "string",
                    "minLength": 1
                },
                "precio": {
                    "title": "Precio",
                    "type": "string",
                    "format": "decimal"
                },
                "stock": {
                    "title": "Stock",
                    "type": "integer",
                    "maximum": 9223372036854775807,
                    "minimum": 0
                },
                "es_basico": {
                    "title": "Es basico",
                    "description": "¿Es un producto de necesidad básica?",
                    "type": "boolean"
                },
                "categoria": {
                    "title": "Categoria",
                    "description": "Categoría del producto (por ejemplo: alimentos, bebidas, aseo, dulces, ferretería, etc.)",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "imagen1": {
                    "title": "Imagen1",
                    "type": "string",
                    "readOnly": true,
                    "x-nullable": true,
                    "format": "uri"
                },
                "imagen2": {
                    "title": "Imagen2",
                    "type": "string",
                    "readOnly": true,
                    "x-nullable": true,
                    "format": "uri"
                },
                "imagen3": {
                    "title": "Imagen3",
                    "type": "string",
                    "readOnly": true,
                    "x-nullable": true,
                    "format": "uri"
                },
                "caracteristicas": {
                    "title": "Caracteristicas",
                    "description": "Características detalladas del producto",
                    "type": "string",
                    "x-nullable": true
                },
                "activo": {
                    "title": "Activo",
                    "type": "boolean"
                },
                "fecha_creacion": {
                    "title": "Fecha creacion",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "fecha_actualizacion": {
                    "title": "Fecha actualizacion",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "tienda": {
                    "title": "Tienda",
                    "type": "integer"
                },
                "proveedor": {
                    "title": "Proveedor",
                    "description": "Proveedor responsable de este producto",
                    "type": "integer"
                }
            }
        },
...
    }
}